fix: relocate codex16 validator for CLI test path - No Veteran Left Behind
